Change observer: Send "errored" event on permission revoke
This makes FileSystemAccessObserverObservation send an "errored" event
and destroy itself when its file handle loses its read permission.

This is a reland of crrev.com/c/5609612. The original failed on Mac
because it didn't support change types. This updates it to expect an
"unknown" event instead of a "modified" event when change info isn't
supported.

Bug: 321980366
Include-Ci-Only-Tests: true
Change-Id: I06373950fd85db6cdd04c278ae887c6cca1806ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5633897
Reviewed-by: Daseul Lee <dslee@chromium.org>
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318670}