[sync] Remove iOS-specific codepaths in SyncPrefs
The features are enabled by default on iOS, so there is no need to use
ifdefs: the code is generally simpler to maintain if it is shared
across platforms.

The change is expected to be a no-op in practice.

Theoretically, there could be clients with
kReplaceSyncPromosWithSignInPromos set to disabled via variations, but
this is considered near-impossible because it could only happen if the
variations seed is more than 6 months old. These cases would presumably
run into other more severe issues (independently of this patch) and,
concerning the behavioral change here, it isn't particularly bad either
(bookmark and reading-list sync would be turned off).

Bug: 346918509
Change-Id: I9a90d38e3cb93c230bf07a7ae7923d0fdfc25e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5647886
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1318671}