[sync] Remove legacy iOS bookmarks/RL state migration
The legacy iOS-specific preference has been deprecated about 6 months
ago and ever since migrated (copied) to a more modern equivalent in
syncer::MaybeMigratePrefsForSyncToSigninPart1().

In this patch, the migration code is deleted and the preference fully
deprecated.

In the unlikely case that the migration code is exercised, the resulting
behavior after this patch would be as if the pref's value was false,
i.e. no opt-in existed. This is accurate for the vast majority of
clients anyway, based on available metrics.

For the extremely-unlikely case where the migration code is exercised
AND the client had previously opted into account bookmarks and reading
lists, this opt-in would be "forgotten". This isn't too severe either,
as the user can turn it back on manually.

Bug: 40266203
Change-Id: I1da70c541e5576f64c049d2e2a2fb56336584b0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5644971
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1318676}