media/mojo/stable_vd_service: Use RegisteredMailboxFrameConverter
This changes the VideoDecoder utility process on ChromeOS to use
RegisteredMailboxFrameConverter to create the exported frames for VD
workflows (with VideoDecoderPipeline). Decoding with VDA is unchanged.

The MailboxFrameRegistry that is used by the frame converter is shared
with the StableVideoDecoderService. When the StableVideoDecoderService
receives a STORAGE_OPAQUE frame, it uses the registry to look up the
original NativePixmapFrameResource and create a GpuMemoryBufferHandle
from the ID and DMABUF file descriptors.

Use of the registered mailbox avoids duplication of the file descriptors
when transferring output frames to the StableVideoDecoderService and
avoids use of STORAGE_DMABUFS or STORAGE_GPU_MEMORY_BUFFER VideoFrames.

Bug: b:332382416
Change-Id: Ib94521081468119731aed37db0a17b89c7317dca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5632234
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318718}