[webauthn] Expand support for disabling the UI
Previously, it was possible to disable the GPM PIN related UI while the
enclave was loading. This CL expands that feature to allow disabling the
UI for most sheets:
* Hover buttons now respect the ui_disabled flag.
* All modal buttons except cancel now respect the ui_disabled flag.
* By default, sheets will show the loading indicator if disabled.

It is still possible to cancel out of the dialog even if the UI is
disabled by design.

Bug: 348445885
Change-Id: I0db036d9e37ca6c8398f71eef94f5d43a30cfc58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5646584
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318771}