documentScan: Fix type converters causing a CHECK error
If a user specifies an option type but the value is a different type,
trying to convert from an OptionSettingPtr to a ScannerOption will
crash.  Add some checks to make sure the input value is the expected
type before trying to access the input value.

Additionally, this now performs a check in the frontend to make sure an
option value has the correct type.  If it doesn't, a kWrongType result
is returned to the caller and the option isn't even sent to the backend
to be set.

Note that there are cases where an option can have an empty value, so
that case is still allowed and sent to the backend.

Bug: 346407721
Change-Id: I718d8163896e86979fc2376be127ab84811026ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5631474
Reviewed-by: Yuzhu Shen <yzshen@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Nathan Muggli <nmuggli@google.com>
Cr-Commit-Position: refs/heads/main@{#1318794}