Move Dawn Ozone Image Representation over to Shared Texture Memory
This is done as far as ozone is concerned but the issue with opaque fences still remains. Basically ozone backing works with sync fences
and the external_vk_image_backing doesnt not.

We are able to make this change only because webGPU for LINUX is not
enabled and will not be for the foreseeable future. WebGPU on linux
will work with this change if it uses the external_vk_image_backing
and associated representations.

This is dependent on:
https://dawn-review.googlesource.com/c/dawn/+/194440

Bug: 330385376
Change-Id: I43c9ae924b44eb7aa4207e1b3e750c66026cbd28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5600123
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Peter McNeeley <petermcneeley@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318796}