Set fixed length for the progress view
We decided to always show two media action buttons on each side of the
progress view, and set the timestamp labels to use the same width as
them, so that the progress view length can be fixed no matter the user
is dragging the progress view or not. The media action buttons will be
disabled if they are not supported.

Screenshot: https://screenshot.googleplex.com/ABpGStts2fgpKbJ

Bug: 348041799
Change-Id: I4f6b83c4b68c3019a315164635caea573d71f870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5646652
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Yiren Wang <yrw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318805}