webauthn: always show Incognito warning.
The code previously showed this warning after the trust-this-device
step, but that's not the order in the mocks, so match the mocks and show
it first.

Also, it wasn't shown in some cases, e.g. it wasn't shown if the macOS
Touch ID sheet was shown.

Bug: 349128164
Change-Id: I56b28f2018f28ea68a3f78cd354579ae2625911f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5648899
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318806}