[RasterInducingScroll] Remove DrawScrollingContentsOp::main_scroll_offset
This removes dependency from the op to the scroll offset, to avoid the
need of PaintArtifactCompositor update on scroll.

I added the field when I was not sure if the scroll offsets would be
always available when they are needed. Based on static code analysis,
for all code paths needing the scroll offsets, the callers supply the
scroll offsets. An experimental run of RasterInducingScroll for all
tests didn't trigger the CHECK in DrawScrollingContentsOp::
GetScrollOffset().

Bug: 40517276
Change-Id: Ib32c3b796a8a0e70befb952d0497aaf711c6a62a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5643299
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318850}