Prepare sanitize for launch
This CL only changes the sanitize flag to be enabled by default and
removes it from the launcher. In the initial launch, sanitize is
kickstarted from the settings menu, so opening it from the launcher
will only show the done dialog which should not be permitted.

BUG=b:317898420
TEST=Manual

Change-Id: I8e4e7b95ecfe1675597b0d51576ad2574e9df7ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5649405
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318860}