media/chromeos/MailboxFrameRegistry: Fix mailbox name generator
The previous implementation of the Mailbox name generator relied on a
function scoped static counter. However, this was not implemented as an
atomic or with proper guarding. If there are multiple
MailboxFrameRegistry objects, there would be racy behavior with multiple
writers into `unique_frame_id`.

There is no need to have per-process uniqueness. Just per-registry.

This moves the counter to a private member variable, which is guaranteed
to be guarded by `lock_`.

Bug: b:332382416
Change-Id: I69608650ae7404c65435e1add2e15e1eebd5f040
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645914
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318874}