[PDF Ink Signatures] Use optional for last event point
The last event point seen during ink stroking is used to support
invalidation.  At times there should not be a valid prior point, at the
start of a stroke.  Future support for stroke segments in
https://crrev.com/c/5549003 also means that at the start of each segment
there should not be a valid prior point.

Rather than carry around a possibly stale value that should be ignored
at certain times, adjust `ink_input_last_event_position` field to be an
optional.  This allows for it to be explicitly reset whenever the value
goes stale and should not be used until a new value is saved.

Bug: 335524380
Change-Id: If489e032bde6f11f0d224c94e9ad19ce22b0c97b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5651684
Commit-Queue: Alan Screen <awscreen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318899}