Early out from creating WebContents/navigation when BrowserContext is shutting down
Otherwise it will result in a CHECK failure when creating the renderer
process for the new WebContents/navigation. This CL makes it so that
we early out at the point where it's still possible to do so (i.e. we
don't already expect the function to return a successful creation), for
the most common triggers for the CHECK failures.

Bug: 40274462
Change-Id: I1ace5cd3cecd11f9ec7efd8df09528b8549c9e10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5647990
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318911}