Use APS Launcher Ordering
The default position for apps will be provided by APS when flag
AppPreloadServiceEnableLauncherOrder is on.

default_app_order is changed to use PackageIds rather than app ids,
and it is reduced to be only the 6 web preinstalled apps, and system
apps.

AppListSyncableService will create the initial set of ordinals
when it starts using the values from default_app_order. It will
query to APS and when it receives the result, it will merge with
the initial list.

Bug: b/327058986
Change-Id: I142fad62fe9c4f4dd62f58400285e029d6f9296f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5619639
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318919}