Do not block user removal by external data removal.
The asynchronous handling is introduced to deal with the issue
that wallpaper's removal needs to wait for system salt to be
ready, because it writes the cached data into the local_state
with the account_id for the user as a key.

Now, it's not written if the user is already removed.
This CL simplifies the async structure in ChromeUserManagerImpl
to reduce the dependency.

Bug: 278643115
Test: Tryjob
Change-Id: Ie574373291072474bc24fcee3164502ec7efeb33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5641799
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318978}