Add category resolved filling readiness metric
As non-Chrome account addresses are shared with Chrome, there is need
for a metric that measures how many Chrome users become read for
Autofill (in the sense of the readiness key metric) due to non-Chrome
account addresses. This CL introduces such a metric by creating a
readiness metric split by profile category.
This is very similar to the existing
Autofill.Leipzig.FillingAssistanceCategory metric, which does the same
for filling assistance.

It is implemented by:
- Making FormEventLoggerBase::RecordFillingReadiness() virtual, so the
  AddressFormEventLogger can override it.
- Adding AddressFormEventLogger::profile_categories_available_, which
  tracks the set of the available profile categories. This complements
  the existing AddressFormEventLogger::record_type_count_.
- Repurposing AddressFormEventLogger::set_record_type_count(count)  to
  SetRecordTypeCount(profiles) and adding logic that populates
  profile_categories_available_.
- Renaming CategoryResolvedFillingAssistanceBucket to
  CategoryResolvedKeyMetricBucket, so it can be reused for the readiness
  metric.
- Extracting a function converting from a set of profile categories to
  CategoryResolvedKeyMetricBucket and using it to emit the category
  resolved assistance and readiness metrics.

Bug: n/a
Change-Id: I54899616b663a7b4961163d5bc9d27d9100c5c6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5646498
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1319015}