[@container] Track snapped scroll-state() as ScrollSnapshotClient
Add a SnappedQueryScrollSnapshot client that updates and validates which
two elements are snapped in the x/y direction of a scroller when there
are snapped targets that are queries using scroll-state() container
queries.

Unlike sticky queries, we use a snapshot client per scroller which acts
as a snap container instead of one for each container for container
queries.

The snapshot client is never removed and there are most likely
invalidation issues for changes in snap-* related and overflow
properties. Invalidations test and fixes will be addressed in later
CL.

Bug: 40279568
Change-Id: I0908eb97518fd649d11d4afdf061edd99ec00f1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5596381
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319065}