Remove TODO to enable SwitchesBetweenTransportAndFeature on android
The existing java bindings have assertions in place to not allow another
sign-in and sync if an account is already signed in. We can update it to
only enable sync if the same account is already signed in but with the
ongoing rollouts happening on clank there's little worth in doing that.

Bug: 348588283
Change-Id: I3c7046d4ded5a15dbc281e22cba03dd90c787967
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5652888
Commit-Queue: Ankush Singh <ankushkush@google.com>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Auto-Submit: Ankush Singh <ankushkush@google.com>
Reviewed-by: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1319078}