webview: Add metric for related website count
An Android app can declare that it has a relationship with a website
by using asset links, or by using intent filters.

This could be helpful in measuring how often WebView is used for browsing
this sort of relationship, vs websites the app has no relationship with.

This metric specifically just focuses on if an app says it has a relationship
rather than verifying if the website declares a relationship back as well.

Bug: 348923073
Change-Id: I96bc5cae90e28e5dce4e8263d07c6f74aee92073
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5644898
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Auto-Submit: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319141}