Extension Safety Hub Refactor Pref Migration
This CL deprecates the use of the boolean kPrefAcknowledgeSafetyCheckWarning pref which only stored whether
a safety check warning had been acknowledged by the user.

Instead, we will now rely only on the integer pref, kPrefAcknowledgeSafetyCheckWarningReason, which stores the specific warning that the user acknowledged.

The above changes require a migration logic that will be removed after a few milestones. The logic consists of:
- If only the boolean kPrefAcknowledgeSafetyCheckWarning pref is set,
  it is removed and the integer kPrefAcknowledgeSafetyCheckWarningReason
  pref is set to the current warning reason.
- If both kPrefAcknowledgeSafetyCheckWarning and
  kPrefAcknowledgeSafetyCheckWarningReason are set, the boolean
  pref is removed.

NOTES:
The code in this CL is not active and will be called in follow up CLs.
The setting of the old boolean pref, kPrefAcknowledgeSafetyCheckWarning, will also be removed in a follow up CL.

Bug: 335639105
Change-Id: I36058b52066c4f5a0fef1ab6ac65bc2a92a8f123
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5639261
Reviewed-by: Anunoy Ghosh <anunoy@chromium.org>
Commit-Queue: Adam Psarouthakis <psarouthakis@google.com>
Cr-Commit-Position: refs/heads/main@{#1319195}