[Extensions] Use pending version for CWS update checks.
When checking for updates through Omaha, the previous
implementation always used the currently installed version of an
extension. This meant that across repeated update checks for an
extension with a delayed (not immediately installed) update, the
same update would be downloaded more than once.

This CL addresses that issue by passing the pending version
down to the update code allowing the update checks to only run if
there is a newer version than the most recent download.

Bug: 348286356
Change-Id: Iaba9682fcbafc84e32c903c79403b74d57a08f40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5642050
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Commit-Queue: Oliver Dunk <oliverdunk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319197}