[CodeHealth] Remove RAW_PTR_EXCLUSION for safe_browsing features list
This refactors the list of safe_browsing base::Features whose state is
shown in the chrome://safe-browsing UI, in order to remove the
RAW_PTR_EXCLUSION annotation on this struct. First converts it to a
simple array of base::Feature*, because there is no need to even include
features whose value for the bool is false. Then converts it to a
function scoped local.

Bug: 346693840
Change-Id: If45e74296e298fa55f77704ed0a9d8b106b6fddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645637
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319226}