[PDF Ink Signatures] Skip drawing for pages that are not visible
For pages that are not visible, there is no need to draw the strokes on
that page. Add InkModule::Client::IsPageVisible() to determine page
visibility, and use that to check if pages need to be drawn or not.

For production usage, PdfViewWebPlugin::IsPageVisible() is a
pass-through to PDFiumEngine::IsPageVisible(), which already exists. For
InkModule's unit tests, add set_page_visibility() to FakeClient, which
is similar to set_page_layouts(). Update existing test cases to call
set_page_visibility() so all pages are visible, like they were without
this CL. Then expand InkModuleStrokeTest.DrawRenderTransform to test
what happens when a page is not visible.

Bug: 335524380
Change-Id: Ie3fb57daa90908fd0ca0ed5ddf3206da277dfcf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5650683
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Andy Phan <andyphan@chromium.org>
Reviewed-by: Alan Screen <awscreen@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319237}