Create AllocationTraceRecorder only when required
Previously, the AllocationTraceRecorder was always created. The idea
was to show in the Crashpad report that it was configured but not
enabled due to runtime flags like MTE being disabled.

With the activation now being explicitly included in the Crashpad
report, we now can create the recorder when it is really needed only.

Bug: 348474004
Change-Id: Ib82a7c297a5ad4002ad9a87db577347ea8173541
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5648045
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Richard Townsend <richard.townsend@arm.com>
Cr-Commit-Position: refs/heads/main@{#1319260}