Fix a dangling pointer issue
Did the following things:
1) In the `UnifiedSliderBubbleController`:
    I) create the slider controller first before removing all the
child views. Otherwise if there are some views stored as a raw
pointer in the controller, there could be a dangling pointer issue.
    II) assign the new `slider_view` to the local `slider_view_`.
Before it was assigned to `nullptr` when the slider got updated,
which is not correct.

2) In each controller, DCHECK if the view is created in each
controller with the `DCHECK_IS_ON` flag. There's no need to store
the slider's raw pointer.

Bug: b/342513656
Change-Id: I6643af567c7f5c08d3f8d790b3fd3aa82138ebc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5651774
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319300}