[omnibox][history-embeddings] Support multiple inputs and interruptions
When the user is typing 'abc', we'll call `Start()` with all 3 inputs
'a', 'ab', and 'abc'. The embeddings service will run 'a' and 'abc', but
will skip 'ab' thanks to the scheduler.

Because the search for 'a' completes first, the embeddings provider
would show results for 'a', instead of 'abc'.

This CL fixes this case by checking if the search response is stale
(i.e. a newer `Search()` has been initiated) or not before updating
matches.

This CL also adds the autocomplete controller as a provider listener for
the embeddings provider. This has to be done for all async providers,
but we forgot to for this provider. Without this, the controller would
only show embeddings suggestions if by luck some other async provider
finishes after the embeddings provider.

Also expands tests.

Bug: 332360110
Change-Id: I34ab008a7eed16374475c01cb59f972f504a624d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5651726
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1319361}