Fix some signed left-shift overflows in libusb
This partially imports
https://github.com/libusb/libusb/commit/8a05a3f4e7424eb2071e4dbe1796cff3b660a141,
but only partially as that patch is very large and does not apply
cleanly.

Specifically, it is UB to shift into the sign bit in a signed integer in
C. So 1 << 31 is not allowed, but 1u << 31 is. This is absurd, so C++20
fixes this. However, libusb builds as C, not C++, and C has not yet
imported this fix, so UBSan flags it.

Example UBSan error:
https://ci.chromium.org/ui/labs/p/chromium/inv/build-8744206501345611617/test/ninja%3A%2F%2Fchrome%2Ftest%3Abrowser_tests%2FActivityLogApiTest.TriggerEvent/variant/720638a27df28011?resultIndex=0

Bug: 40248746
Change-Id: I52f29045d8fccf41e8c6f565e317b98417894d12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5647608
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319362}